Added EC2ClientFactory
